# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FauxVault API
  description: |
    REST API for FauxVault — a purposefully vulnerable web banking application
    for CS467 (Spring 2026). Demonstrates OWASP Top 10 attack vectors with
    toggleable vulnerable/hardened states.

    ## Endpoint Status
    - **Implemented** — working and tested against PostgreSQL
    - **Planned** — not yet built; included so the frontend team knows the
      target contract

    ## Authentication
    Protected endpoints require a Bearer JWT in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    Tokens are returned by the login and register endpoints.

    ## Error Format
    All errors follow a consistent shape:
    ```json
    {
      "error": {
        "status": 400,
        "message": "Human-readable message",
        "code": "MACHINE_READABLE_CODE"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: FauxVault Team
servers:
  - url: http://localhost:3001/api
    description: Local development (host machine)
  - url: http://localhost:80/api
    description: Docker Compose
...
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: "[PLANNED] Get the authenticated user's profile"
      x-status: planned
      description: |
        **Status: PLANNED — not yet implemented**

        Returns profile info for the authenticated user. Needed by:
        - Frontend `users.js` `getProfile()`
        - Darrin's Mass Assignment module (API3) — in vulnerable mode this
          endpoint returns ALL fields including sensitive data; in hardened
          mode it returns filtered fields only.

        Planned response shape based on frontend mock and DB schema:
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      username:
                        type: string
                      email:
                        type: string
                        format: email
                      name:
                        type: string
                      role:
                        type: string
                      accountNumber:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
              example:
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: scurry
                  email: steph.curry@email.com
                  name: Steph Curry
                  role: user
                  accountNumber: FAUX-M3KQ8P1Z
                  createdAt: '2026-04-27T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Users]
      summary: "[PLANNED] Update the authenticated user's profile"
      x-status: planned
      description: |
        **Status: PLANNED — not yet implemented**

        Updates profile fields for the authenticated user. Needed by:
        - Frontend `users.js` `updateProfile(data)`
        - Darrin's Mass Assignment module (API3) — in vulnerable mode this
          endpoint accepts ANY field including `role` and `isAdmin`; in
          hardened mode it only accepts whitelisted fields (name, email).
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
            example:
              name: Stephen Curry
              email: steph.curry@email.com
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      username:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                      role:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─────────────────────────────────────────────
  # Health
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by /auth/login or /auth/register
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            status:
              type: integer
            message:
              type: string
            code:
              type: string
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            no_token:
              value:
                error:
                  status: 401
                  message: no token provided
                  code: AUTH_REQUIRED
            invalid_token:
              value:
                error:
                  status: 401
                  message: Invalid or expired token
                  code: INVALID_TOKEN